reordered schema creation, moved types and errors to appropriate files, changed post to put
