+++ src/db/postgres.ts
@@ -21,6 +21,12 @@ export const initializePostgres = async () => {
         created_at TIMESTAMP DEFAULT NOW()
       );
 
+      CREATE TABLE IF NOT EXISTS queues (
+        id          VARCHAR(50) PRIMARY KEY,
+        name        TEXT        NOT NULL,
+        created_at  TIMESTAMP   DEFAULT NOW()
+      );
+
       CREATE TABLE IF NOT EXISTS root_spans (
         id VARCHAR(50) PRIMARY KEY,
         trace_id VARCHAR(50) NOT NULL,
@@ -51,12 +57,6 @@ export const initializePostgres = async () => {
         id VARCHAR(50) PRIMARY KEY,
         annotation_id VARCHAR(50) NOT NULL REFERENCES annotations(id) ON DELETE CASCADE,
         category_id VARCHAR(50) NOT NULL REFERENCES categories(id) ON DELETE CASCADE
-      );
-
-      CREATE TABLE IF NOT EXISTS queues (
-        id          VARCHAR(50) PRIMARY KEY,
-        name        TEXT        NOT NULL,
-        created_at  TIMESTAMP   DEFAULT NOW()
       );
 
     `);
